refactor: type, interface 정리 #22
