+++ src/data/chemical.ts
@@ -1,18 +1,4 @@
-interface IChemicalData {
-  Henry: number | null
-  logKoclogKd: number | null
-  'Koc/Kd': 'Koc' | 'Kd' | null
-  Sfo: number | null
-  RfDo: number | null
-  RfDd: number | null
-  SFd: number | null
-  URF: number | null
-  Rfc: number | null
-  Dair: number | null
-  Dwat: number | null
-  BCF: number | null
-  IUR: number | null
-}
+import { IChemicalData } from '../types/chemical.type'
 
 export const CHEMICAL_DATABASE: Record<string, IChemicalData> = {
   Acenaphthene: {
