+++ src/data/exposure.ts
@@ -1,18 +1,4 @@
-interface IExposureData {
-  Setting: 'Residential' | 'Commercial'
-  DepthBelowGradeToBaseOfFoundation: number
-  FoundationThickness: number
-  FractionOfFoundationAreaWithCracks: number
-  EnclosedSpaceFloorArea: number
-  EnclosedSpaceHeight: number
-  IndoorAirExchangeRate: number
-  BuildingVentilationRate: number
-  QsoilQbuilding: number
-  AverageVaporFlowRateIntoBldg: number | null
-  eta: number | null
-  Abf: number
-  Hb: number
-}
+import { IExposureData } from '../types/exposure.type'
 
 export const EXPOSURE_DATABASE: Record<string, IExposureData> = {
   'Closed crawl space w/ slab': {
