+++ src/data/soil.ts
@@ -1,18 +1,4 @@
-interface ISoilData {
-  Symbol: string
-  TotalPorosity: number | null
-  WaterFilledPorosoty: number | null
-  BulkDensity: number | null
-  SaturatedHydraulicConductivity: number | null
-  CapillaryZoneWaterFilledPorosity: number | null
-  CapillaryZoneHeight: number | null
-  a1: number | null
-  N: number | null
-  M: number | null
-  qr: number | null
-  MeanGrainDiameter: number | null
-  nwRange: number | null
-}
+import { ISoilData } from '../types/soil.type'
 
 export const SOIL_DATABASE: Record<string, ISoilData> = {
   Clay: {
