+++ src/types/chemical.type.ts
@@ -1,0 +1,15 @@
+export interface IChemicalData {
+  Henry: number | null
+  logKoclogKd: number | null
+  'Koc/Kd': 'Koc' | 'Kd' | null
+  Sfo: number | null
+  RfDo: number | null
+  RfDd: number | null
+  SFd: number | null
+  URF: number | null
+  Rfc: number | null
+  Dair: number | null
+  Dwat: number | null
+  BCF: number | null
+  IUR: number | null
+}
