+++ src/types/soil.type.ts
@@ -1,0 +1,15 @@
+export interface ISoilData {
+  Symbol: string
+  TotalPorosity: number | null
+  WaterFilledPorosoty: number | null
+  BulkDensity: number | null
+  SaturatedHydraulicConductivity: number | null
+  CapillaryZoneWaterFilledPorosity: number | null
+  CapillaryZoneHeight: number | null
+  a1: number | null
+  N: number | null
+  M: number | null
+  qr: number | null
+  MeanGrainDiameter: number | null
+  nwRange: number | null
+}
